Add generate workflow
This workflow for the `grano` CLI allows for generating potentially useful graph sequence examples. In particular, it allows for generating constant and dynamic random graph sequences of varying lengths and varying sizes. By outputting results as a `scaffold.json` file, these sequences can be used as input to a Parse Workflow.

Closes #12